Move geograpic collector triggers to a sub-package
