+++ pytroll_collectors/triggers/__init__.py
@@ -1,0 +1,38 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) 2012, 2014, 2015, 2019 Martin Raspaud
+#
+# Author(s):
+#
+#   Kristian Rune Larsen <krl@dmi.dk>
+#   Martin Raspaud <martin.raspaud@smhi.se>
+#   Panu Lahtinen <panu.lahtinen@fmi.fi>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Triggers for region_collectors."""
+
+import logging
+
+from ._inotify import InotifyTrigger
+from ._posttroll import PostTrollTrigger
+
+logger = logging.getLogger(__name__)
+
+try:
+    from ._watchdog import WatchDogTrigger
+except ImportError:
+    logger.exception("Watchdog import failed!")
+    WatchDogTrigger = None
